fixed divide by zero bug in fminunc by fixing alt type
when alt is passed to octave as a float, it's kept that
way internallly in octave. this makes the gradient calculation
in fminunc fail with a divide by zero error

+++ src/mvp/MVPJobBase.h
@@ -39,7 +39,7 @@ struct MVPAlgorithmVar {
 
   ::octave_scalar_map to_octave() const {
     ::octave_scalar_map result;
-    result.setfield("alt", alt);
+    result.setfield("alt", double(alt));
     result.setfield("orientation", vw::octave::vector_to_octave(orientation));
     result.setfield("windows", vw::octave::vector_to_octave(windows));
     return result;
